Move hwcomposer to separate file.
