+++ asteroidos/Makefile
@@ -9,7 +9,7 @@
 
 OBJS      = asteroid.o main.o cpu.o memory.o video.o input.o sound.o gui.o \
             cheats.o zip.o  arm_stub.o warm.o cpu_threaded.o\
-	    gles_video.o video_blend.o
+	    hwcomposer.o gles_video.o video_blend.o
 
 BIN       = gpsp
