+++ asteroidos/gles_video.h
@@ -1,14 +1,4 @@
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-
 void video_init(uint32_t width,uint32_t height, uint32_t f);
 void video_close();
 void video_draw(uint16_t *pixels);
 void video_set_filter(uint32_t filter);
-#ifdef __cplusplus
-}
-#endif
