+++ asteroidos/hwcomposer.cpp
@@ -1,0 +1,262 @@
+#include "GLES/gl.h"
+#include "EGL/egl.h"
+#include "EGL/eglext.h"
+#include "GLES2/gl2.h"
+#include <assert.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <memory.h>
+#include <unistd.h>
+#include <hwcomposer_window.h>
+#include <hardware/hardware.h>
+#include <hardware/hwcomposer.h>
+#include <malloc.h>
+#include <sync/sync.h>
+
+#include "hwcomposer.h"
+
+class HWComposer : public HWComposerNativeWindow
+{
+	private:
+		hwc_layer_1_t *fblayer;
+		hwc_composer_device_1_t *hwcdevice;
+		hwc_display_contents_1_t **mlist;
+	protected:
+		void present(HWComposerNativeWindowBuffer *buffer);
+
+	public:
+
+		HWComposer(unsigned int width, unsigned int height, unsigned int format, hwc_composer_device_1_t *device, hwc_display_contents_1_t **mList, hwc_layer_1_t *layer);
+		void set();	
+};
+
+HWComposer::HWComposer(unsigned int width, unsigned int height, unsigned int format, hwc_composer_device_1_t *device, hwc_display_contents_1_t **mList, hwc_layer_1_t *layer) : HWComposerNativeWindow(width, height, format)
+{
+	fblayer = layer;
+	hwcdevice = device;
+	mlist = mList;
+}
+
+void HWComposer::present(HWComposerNativeWindowBuffer *buffer)
+{
+	int oldretire = mlist[0]->retireFenceFd;
+	mlist[0]->retireFenceFd = -1;
+	fblayer->handle = buffer->handle;
+	fblayer->acquireFenceFd = getFenceBufferFd(buffer);
+	fblayer->releaseFenceFd = -1;
+	int err = hwcdevice->prepare(hwcdevice, HWC_NUM_DISPLAY_TYPES, mlist);
+	assert(err == 0);
+
+	err = hwcdevice->set(hwcdevice, HWC_NUM_DISPLAY_TYPES, mlist);
+	// in android surfaceflinger ignores the return value as not all display types may be supported
+	setFenceBufferFd(buffer, fblayer->releaseFenceFd);
+
+	if (oldretire != -1)
+	{   
+		//sync_wait(oldretire, -1);
+		close(oldretire);
+	}
+} 
+
+inline static uint32_t interpreted_version(hw_device_t *hwc_device)
+{
+	uint32_t version = hwc_device->version;
+
+	if ((version & 0xffff0000) == 0) {
+		// Assume header version is always 1
+		uint32_t header_version = 1;
+
+		// Legacy version encoding
+		version = (version << 16) | header_version;
+	}
+	return version;
+}
+
+uint32_t screen_width = 0;
+uint32_t screen_height = 0;
+
+EGLDisplay display = NULL;
+EGLSurface surface = NULL;
+EGLContext context = NULL;
+static HWComposer *win = NULL;
+
+static GLfloat proj[4][4];
+static GLint filter_min;
+static GLint filter_mag;
+
+void hwcomposer_init() {
+	EGLConfig ecfg;
+	EGLint num_config;
+	EGLint attr[] = {       // some attributes to set up our egl-interface
+		EGL_BUFFER_SIZE, 32,
+		EGL_RENDERABLE_TYPE,
+		EGL_OPENGL_ES2_BIT,
+		EGL_NONE
+	};
+	EGLint ctxattr[] = {
+		EGL_CONTEXT_CLIENT_VERSION, 2,
+		EGL_NONE
+	};
+
+	EGLBoolean rv;
+
+	int err;	
+
+        hw_module_t const* module = NULL;
+        err = hw_get_module(GRALLOC_HARDWARE_MODULE_ID, &module);
+        assert(err == 0);
+        //framebuffer_device_t* fbDev = NULL;
+        //framebuffer_open(module, &fbDev);
+
+	hw_module_t *hwcModule = 0;
+	hwc_composer_device_1_t *hwcDevicePtr = 0;
+
+	err = hw_get_module(HWC_HARDWARE_MODULE_ID, (const hw_module_t **) &hwcModule);
+	assert(err == 0);
+	
+	err = hwc_open_1(hwcModule, &hwcDevicePtr);
+	assert(err == 0);
+
+	hw_device_t *hwcDevice = &hwcDevicePtr->common;
+
+	uint32_t hwc_version = interpreted_version(hwcDevice);
+
+#ifdef HWC_DEVICE_API_VERSION_1_4
+	if (hwc_version == HWC_DEVICE_API_VERSION_1_4) {
+		hwcDevicePtr->setPowerMode(hwcDevicePtr, 0, HWC_POWER_MODE_NORMAL);
+	} else
+#endif
+#ifdef HWC_DEVICE_API_VERSION_1_5
+	if (hwc_version == HWC_DEVICE_API_VERSION_1_5) {
+		hwcDevicePtr->setPowerMode(hwcDevicePtr, 0, HWC_POWER_MODE_NORMAL);
+	} else
+#endif
+		hwcDevicePtr->blank(hwcDevicePtr, 0, 0);
+
+	uint32_t configs[5];
+	size_t numConfigs = 5;
+
+	err = hwcDevicePtr->getDisplayConfigs(hwcDevicePtr, 0, configs, &numConfigs);
+	assert (err == 0);
+
+	int32_t attr_values[2];
+	uint32_t attributes[] = { HWC_DISPLAY_WIDTH, HWC_DISPLAY_HEIGHT, HWC_DISPLAY_NO_ATTRIBUTE }; 
+
+	hwcDevicePtr->getDisplayAttributes(hwcDevicePtr, 0,
+			configs[0], attributes, attr_values);
+
+	printf("width: %i height: %i\n", attr_values[0], attr_values[1]);
+
+	size_t size = sizeof(hwc_display_contents_1_t) + 2 * sizeof(hwc_layer_1_t);
+	hwc_display_contents_1_t *list = (hwc_display_contents_1_t *) malloc(size);
+	hwc_display_contents_1_t **mList = (hwc_display_contents_1_t **) malloc(HWC_NUM_DISPLAY_TYPES * sizeof(hwc_display_contents_1_t *));
+	const hwc_rect_t r = { 0, 0, attr_values[0], attr_values[1] };
+
+	int counter = 0;
+	for (; counter < HWC_NUM_DISPLAY_TYPES; counter++)
+		mList[counter] = NULL;
+
+	// Assign the layer list only to the first display,
+	// otherwise HWC might freeze if others are disconnected
+	mList[0] = list;
+
+	hwc_layer_1_t *layer = &list->hwLayers[0];
+	memset(layer, 0, sizeof(hwc_layer_1_t));
+	layer->compositionType = HWC_FRAMEBUFFER;
+	layer->hints = 0;
+	layer->flags = 0;
+	layer->handle = 0;
+	layer->transform = 0;
+	layer->blending = HWC_BLENDING_NONE;
+#ifdef HWC_DEVICE_API_VERSION_1_3
+	layer->sourceCropf.top = 0.0f;
+	layer->sourceCropf.left = 0.0f;
+	layer->sourceCropf.bottom = (float) attr_values[1];
+	layer->sourceCropf.right = (float) attr_values[0];
+#else
+	layer->sourceCrop = r;
+#endif
+	layer->displayFrame = r;
+	layer->visibleRegionScreen.numRects = 1;
+	layer->visibleRegionScreen.rects = &layer->displayFrame;
+	layer->acquireFenceFd = -1;
+	layer->releaseFenceFd = -1;
+#if (ANDROID_VERSION_MAJOR >= 4) && (ANDROID_VERSION_MINOR >= 3) || (ANDROID_VERSION_MAJOR >= 5)
+	// We've observed that qualcomm chipsets enters into compositionType == 6
+	// (HWC_BLIT), an undocumented composition type which gives us rendering
+	// glitches and warnings in logcat. By setting the planarAlpha to non-
+	// opaque, we attempt to force the HWC into using HWC_FRAMEBUFFER for this
+	// layer so the HWC_FRAMEBUFFER_TARGET layer actually gets used.
+	bool tryToForceGLES = getenv("QPA_HWC_FORCE_GLES") != NULL;
+	layer->planeAlpha = tryToForceGLES ? 1 : 255;
+#endif
+#ifdef HWC_DEVICE_API_VERSION_1_5
+	layer->surfaceDamage.numRects = 0;
+#endif
+
+	layer = &list->hwLayers[1];
+	memset(layer, 0, sizeof(hwc_layer_1_t));
+	layer->compositionType = HWC_FRAMEBUFFER_TARGET;
+	layer->hints = 0;
+	layer->flags = 0;
+	layer->handle = 0;
+	layer->transform = 0;
+	layer->blending = HWC_BLENDING_NONE;
+#ifdef HWC_DEVICE_API_VERSION_1_3
+	layer->sourceCropf.top = 0.0f;
+	layer->sourceCropf.left = 0.0f;
+	layer->sourceCropf.bottom = (float) attr_values[1];
+	layer->sourceCropf.right = (float) attr_values[0];
+#else
+	layer->sourceCrop = r;
+#endif
+	layer->displayFrame = r;
+	layer->visibleRegionScreen.numRects = 1;
+	layer->visibleRegionScreen.rects = &layer->displayFrame;
+	layer->acquireFenceFd = -1;
+	layer->releaseFenceFd = -1;
+#if (ANDROID_VERSION_MAJOR >= 4) && (ANDROID_VERSION_MINOR >= 3) || (ANDROID_VERSION_MAJOR >= 5)
+	layer->planeAlpha = 0xff;
+#endif
+#ifdef HWC_DEVICE_API_VERSION_1_5
+	layer->surfaceDamage.numRects = 0;
+#endif
+
+	list->retireFenceFd = -1;
+	list->flags = HWC_GEOMETRY_CHANGED;
+	list->numHwLayers = 2;
+
+
+	screen_width = attr_values[0];
+	screen_height = attr_values[1];
+
+	win = new HWComposer(attr_values[0], attr_values[1], HAL_PIXEL_FORMAT_RGBA_8888, hwcDevicePtr, mList, &list->hwLayers[1]);
+
+	display = eglGetDisplay(NULL);
+	assert(eglGetError() == EGL_SUCCESS);
+	assert(display != EGL_NO_DISPLAY);
+
+	rv = eglInitialize(display, 0, 0);
+	assert(eglGetError() == EGL_SUCCESS);
+	assert(rv == EGL_TRUE);
+
+	eglChooseConfig((EGLDisplay) display, attr, &ecfg, 1, &num_config);
+	assert(eglGetError() == EGL_SUCCESS);
+	assert(rv == EGL_TRUE);
+
+
+
+	surface = eglCreateWindowSurface((EGLDisplay) display, ecfg, (EGLNativeWindowType) static_cast<ANativeWindow *> (win), NULL);
+	assert(eglGetError() == EGL_SUCCESS);
+	assert(surface != EGL_NO_SURFACE);
+
+	context = eglCreateContext((EGLDisplay) display, ecfg, EGL_NO_CONTEXT, ctxattr);
+	assert(eglGetError() == EGL_SUCCESS);
+	assert(context != EGL_NO_CONTEXT);
+
+	assert(eglMakeCurrent((EGLDisplay) display, surface, surface, context) == EGL_TRUE);
+
+	const char *version = (const char *)glGetString(GL_VERSION);
+	assert(version);
+	printf("%s\n",version);
+}
